Extract PositionCard for reuse across Earn
Fixed term Earn details renders the same card, so both pages now
use one component, and its label moves to pages.earn.stats.

Also add the missing rule under the divider and swap the empty
grid column for a gap.
